fix(core): strictly enforce jsx2ttl as single JSX implementation
